[feat] post like api 사용 추가 - post like api 추가 - like 정보를 바탕으로 초기 like 값 설정   - state 중복으로 부모로 끌어올림 [bugfix] like + share 버튼 layout 수정 - layout height, 정렬 오류 - layout height 100%로 수정, 정렬 flext로 display 설정 후 center로 배치
